refactor(gateway): split resolvers monolith into domain modules
