+++ src/gateway/graphql/resolvers/utilities.ts
@@ -1,0 +1,14 @@
+import { parseUserAgent } from '@/gateway/services/user-agent'
+import { lookupIp } from '@/gateway/services/geolocation'
+
+export const utilitiesResolvers = {
+  Query: {
+    parseUserAgent: (_root: unknown, args: { userAgent: string }) => {
+      return parseUserAgent(args.userAgent)
+    },
+
+    geolocateIP: (_root: unknown, args: { ip: string }) => {
+      return lookupIp(args.ip)
+    },
+  },
+}
